Update material-top-tab-navigator.md
There is no easy way to center the indicator or any supporting documentation. I suggest a re-write of the component to not use Absolute for positioning but this will do for now in helping the majority of people (As seen in the official Discord channel who could still not solve this).