Add reminder to include super(props)
The example doesn't include this dependency and people will have to remember to add it on their own. It would be useful to have a fully working Button class in the example. I pasted the example, forgot to add the constructor function on my first go around and only remembered to add it in after getting the error again.