Add warning about remote debugging
Deep-linking does not function as expected when the remote debugger is enabled. If the app is already running on the phone (foreground or background) it works as expected, but if the app is closed opening a link will only ever open the initial route. This is an annoying and opaque issue to diagnose and it would be a boon to developers to be warned about it in the documentation.