Add warning about remote debugger
If remote debugger is enabled deep-linking does not function as expected. If the app is in a closed state (not running in foreground or background) links will only ever open the initial route. This is an opaque problem that is difficult to diagnose. A warning in the docs would be a boon to developers.