Update navigation prop setOptions example
Not entirely sure if this is intended, but I believe in this example use of `React.useLayoutEffect` is misleading.
In the example component no specific DOM manipulations/calculations are made, so `React.useEffect` should be just enough for this case. 

I've tried using `React.useEffect` and it works fine for updating the header title.