docs: Update 'Passing params to a previous screen'
I'm currently developing app that passes params from subsequent screens. One screen also has a screen that passes back additional route params.
I was not sure if I needed to pass all params to new screen or if the behaviour is the same as ```setParams```.
It turns out that second behaviour is present.
I think it will be useful to mention it in the documentation.