docs: mention Settings would be the route an app would be displaying for listed JS state
Further flesh out which route an app is displayed that would correspond
to the navigation state represented by the listed JavaScript object. For
example, it might not be immediately obvious to a newcomer that among
the 2 routes of `Home` and `Settings`, it would be the `Settings` route
that is displayed, and not `Home`.

Signed-off-by: Ray Chuan Tay <rtay@vmware.com>